chore: Move show options form from bottom sheet to dedicated panel
Make things more friendly for TV ui, and less prone to clicking outside and losing form data on other devices.
